refactor(op): split lib.rs into per-concern modules
